refactor : 이미지 도메인 클래스 이동
