+++ src/main/java/nextstep/courses/domain/Session.java
@@ -1,6 +1,6 @@
 package nextstep.courses.domain;
 
-import nextstep.common.domain.Image;
+import nextstep.image.domain.Image;
 import nextstep.courses.exception.ExceededStudentCount;
 import nextstep.courses.exception.OutOfRegistrationPeriod;
 import nextstep.utils.CommunicationTerm;
