+++ src/main/java/nextstep/image/infrastructure/JdbcImageRepository.java
@@ -1,7 +1,7 @@
-package nextstep.common.infrastructure;
+package nextstep.image.infrastructure;
 
-import nextstep.common.domain.Image;
-import nextstep.common.domain.ImageRepository;
+import nextstep.image.domain.Image;
+import nextstep.image.domain.ImageRepository;
 import org.springframework.jdbc.core.JdbcTemplate;
 import org.springframework.jdbc.core.RowMapper;
 import org.springframework.jdbc.core.namedparam.MapSqlParameterSource;
